refactor: 部分文件安装为hi调整至项目目录

+++ tools/binutils-gdb.sh
@@ -11,28 +11,34 @@ notice() {
 fail() {
     echo -e "\033[41;37m 失败 \033[0m $1"
 }
+max_thread=$(cat /proc/cpuinfo| grep "processor"| wc -l)
+export JOBS=$max_thread
+output_dir="$root_dir/output"
+toolchain_dir="$output_dir/toolchain"
 
 if [ ! -d "$root_dir/binutils-gdb" ]; then
   notice "binutils-gdb not cloned, start to clone......"
   git clone https://sourceware.org/git/binutils-gdb.git
 fi
 
-max_thread=$(cat /proc/cpuinfo| grep "processor"| wc -l)
-export JOBS=$max_thread
-
-cd binutils-gdb
-mkdir build
-cd build
-toolchain_dir="$root_dir/toolchain"
+project_dir="$root_dir/binutils-gdb"
+build_dir="$project_dir/build"
+mkdir -p "$build_dir"
 mkdir -p "$toolchain_dir"
+cd binutils-gdb/build
 
-notice "configure binutils-gdb for x86_64"
-../configure --prefix=$toolchain_dir --target=x86_64-linux-gnu --disable-nls --enable-shared --enable-64-bit-bfd --disable-werror
-notice "make binutils-gdb for x86_64"
-make && make install
+if [ ! -f "$toolchain_dir/bin/x86_64-linux-gnu-ld" ];then
+  notice "configure binutils-gdb for x86_64"
+  rm ./* -rf
+  ../configure --prefix=$toolchain_dir --target=x86_64-linux-gnu --disable-nls --enable-shared --enable-64-bit-bfd --disable-werror
+  notice "make binutils-gdb for x86_64"
+  make -j$max_thread && make install
+fi
 
-rm ./* -rf
-notice "configure binutils-gdb for loongarch64"
-../configure --prefix=$toolchain_dir --target=loongarch64-linux-gnu --with-sysroot --disable-nls --enable-shared --enable-64-bit-bfd --disable-werror
-notice "make binutils-gdb for loongarch64"
-make && make install
+if [ ! -f "$toolchain_dir/bin/loongarch64-linux-gnu-ld" ];then
+  rm ./* -rf
+  notice "configure binutils-gdb for loongarch64"
+  ../configure --prefix=$toolchain_dir --target=loongarch64-linux-gnu --with-sysroot --disable-nls --enable-shared --enable-64-bit-bfd --disable-werror
+  notice "make binutils-gdb for loongarch64"
+  make -j$max_thread && make install
+fi
